M2-6953: Code review fixes
